requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.10
Django==3.2.7
django-cors-headers==3.8.0
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
idna==3.3
imagezmq==1.1.1
numpy==1.21.2
opencv-contrib-python==4.5.5.62
pandas==1.3.3
Pillow==8.3.2
PyJWT==2.3.0
python-dateutil==2.8.2
pytz==2021.3
pyzmq==22.3.0
requests==2.26.0
rest-pandas==1.1.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.8
